rename: DTO이름 수정, 필요없는 어노테이션 제거
