+++ src/main/java/com/potatocake/everymoment/dto/request/CategoryRequest.java
@@ -1,0 +1,8 @@
+package com.potatocake.everymoment.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class CategoryRequest {
+    private Long categoryId;
+}
