--- src/main/java/com/potatocake/everymoment/dto/request/CategoryRequestDTO.java
@@ -1,12 +1,0 @@
-package com.potatocake.everymoment.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CategoryRequestDTO {
-    private Long categoryId;
-}
