+++ src/main/java/com/potatocake/everymoment/dto/request/FileRequest.java
@@ -1,0 +1,9 @@
+package com.potatocake.everymoment.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class FileRequest {
+    private String imageUrl;
+    private int order;
+}
