+++ src/main/java/com/potatocake/everymoment/dto/response/CategoryResponse.java
@@ -1,15 +1,11 @@
 package com.potatocake.everymoment.dto.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
-@AllArgsConstructor
 @Builder
-public class CategoryResponseDTO {
+public class CategoryResponse {
     private long id;
     private String categoryName;
 }
