+++ src/main/java/com/potatocake/everymoment/dto/response/FileResponse.java
@@ -1,15 +1,12 @@
 package com.potatocake.everymoment.dto.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
-@AllArgsConstructor
 @Builder
-public class ThumbnailResponseDTO {
+public class FileResponse {
     private Long id;
     private String imageUrl;
+    private int order;
 }
