+++ src/main/java/com/potatocake/everymoment/dto/response/ThumbnailResponse.java
@@ -1,16 +1,11 @@
 package com.potatocake.everymoment.dto.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
-@AllArgsConstructor
 @Builder
-public class FileResponseDTO {
+public class ThumbnailResponse {
     private Long id;
     private String imageUrl;
-    private int order;
 }
